+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">qiitaApplication</string>
-    <string name="title_article_activity">Qiita記事一覧</string>
+    <string name="title_article_activity">Qiita</string>
     <string name="label_user_name">ユーザ名：</string>
     <string name="label_likes_count">いいね：</string>
     <string name="label_created_at">作成日：</string>
@@ -8,4 +8,7 @@
     <string name="button_show_all_record">保存レコードのログを出す</string>
     <string name="title_favorite_activity">お気に入り一覧</string>
     <string name="button_favorite_activity">お気に入り一覧</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
